Refactor Correlation ID handling in middleware tests
Introduce a new test to ensure correlation ID is used as-is when present in the header. Modify an existing test to check the correlation ID in the response header using the default ToString() method. Remove GuidDigitsFormat constant and update code to use default ToString() method for generating new correlation IDs.